feat: 쿠폰 스케쥴러 레이어 위치 변경
