+++ src/main/java/kr/hhplus/be/server/presentation/coupon/CouponScheduler.java
@@ -1,0 +1,16 @@
+package kr.hhplus.be.server.presentation.coupon;
+
+import kr.hhplus.be.server.domain.coupon.CouponService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@RequiredArgsConstructor
+@Component
+public class CouponScheduler {
+
+    private final CouponService couponService;
+
+    public void expireCoupons() {
+        couponService.expireCoupons();
+    }
+}
